+++ src/Parser/XMLDaoParser.php
@@ -11,8 +11,6 @@
 namespace Jelix\Dao\Parser;
 
 use Jelix\Dao\ContextInterface;
-use Jelix\Dao\ContextInterface2;
-use Jelix\Dao\CustomClassFile;
 use Jelix\Dao\CustomClassFileInterface;
 use Jelix\Dao\DaoFileInterface;
 
@@ -280,12 +278,7 @@ class XMLDaoParser
         if (isset($xml->factory)) {
 
             if (isset($xml->factory[0]['extends'])) {
-                if ($this->context instanceof ContextInterface2) {
-                    $this->_customFactory =  $this->context->resolveCustomFactoryClassPath((string) $xml->factory[0]['extends']);
-                }
-                else {
-                    $this->_customFactory = new CustomClassFile((string) $xml->factory[0]['extends']);
-                }
+                $this->_customFactory =  $this->context->resolveCustomFactoryClassPath((string) $xml->factory[0]['extends']);
             }
 
             if (isset($xml->factory[0]['events'])) {